requirements.txt
ansible==2.9.2
attrs==19.3.0
autopep8==1.4.4
bitmex==0.2.2
bitmex-ws==0.4.0
bravado==10.4.3
bravado-core==5.16.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
idna==2.9
importlib-metadata==1.5.0
Jinja2==2.11.1
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
MarkupSafe==1.1.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
pycodestyle==2.5.0
pycparser==2.19
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rfc3987==1.3.8
simplejson==3.17.0
six==1.13.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
typing-extensions==3.7.4.1
urllib3==1.25.8
webcolors==1.11.1
websocket-client==0.53.0
zipp==3.0.0
